+++ libqpdf/QPDF_Stream.cc
@@ -466,7 +466,7 @@ QPDF_Stream::filterable(std::vector<std::string>& filters,
 
 bool
 QPDF_Stream::pipeStreamData(Pipeline* pipeline,
-                            unsigned long encode_flags,
+                            int encode_flags,
                             qpdf_stream_decode_level_e decode_level,
                             bool suppress_warnings, bool will_retry)
 {
